refactor: expose shared Firestore API from one module
