+++ src/features/sign-in/model/signIn.spec.ts
@@ -6,7 +6,7 @@ const mocks = vi.hoisted(() => ({
   auth: { currentUser: null as { isAnonymous: boolean } | null },
 }));
 
-vi.mock("@/shared/firebase", () => ({ auth: mocks.auth }));
+vi.mock("@/shared/lib/firestoreRuntime", () => ({ auth: mocks.auth }));
 vi.mock("firebase/auth");
 
 import { loginGoogle } from "./signIn";
