+++ src/features/sign-out/model/signOut.ts
@@ -1,5 +1,5 @@
 import { signOut } from "firebase/auth";
 
-import { auth } from "@/shared/firebase";
+import { auth } from "@/shared/firestore";
 
 export const signOutCurrentUser = (): Promise<void> => signOut(auth);
