+++ src/pages/deck-list/ui/DeckListPage.spec.tsx
@@ -94,7 +94,7 @@ vi.mock("@/features/deck-list", () => ({
   },
 }));
 vi.mock("react-router-dom", () => ({ useNavigate: () => mocks.navigate }));
-vi.mock("@/shared/firebase", () => ({ auth: {}, db: {} }));
+vi.mock("@/shared/lib/firestoreRuntime", () => ({ auth: {}, db: {} }));
 
 import { DeckListPage } from "./DeckListPage";
