+++ src/pages/deck-swiper/ui/DeckSwiperPage.spec.tsx
@@ -49,7 +49,7 @@ const mocks = vi.hoisted(() => ({
   setDarkMode: vi.fn(),
 }));
 
-vi.mock("@/shared/firebase", () => ({ auth: {} }));
+vi.mock("@/shared/lib/firestoreRuntime", () => ({ auth: {} }));
 
 vi.mock("@/entities/preferences", () => ({
   usePreferences: () => {
