--- src/shared/firebase.ts
@@ -1,1 +1,0 @@
-export { auth, db } from "./firestore/firebase";
