--- src/shared/firestore/index.ts
@@ -1,7 +1,0 @@
-export {
-  firestoreTimestampDateSchema,
-  getTimestamp,
-  omitUndefined,
-  parseFirestoreDocument,
-} from "./firestoreDocument";
-export { toRemoteById } from "./remoteSnapshot";
